Write modes with zero eigenvalues in the NMD file
In the method writeNMD in nmdfile.py, add the parameter zeros.
It controls writing zero eigenvale modes to the NMD file. It
can be very useful to inspect rigid-body motion vectors,
especially in the context of protein docking.

If zeros is **False** (by default), this function skips modes
with zero eigenvalues.

If zeros is **True**, modes with zero eigenvalues are
written out, their scaling factor being the square-root
of the inverse of the mode number times 0.0001. This provides
descending factors consistent with the NMA modes.