Create CONTRIBUTING.md
This is a first draft, which is based on the ProDy website page http://prody.csb.pitt.edu/manual/devel/develop.html. Please see that page for updates.

We probably also want to see if we need to update that. I noticed that it says abakan in many places, which we may want to change.